requirements.txt
python-dotenv~=1.0.0
aiogram~=3.0.0
aiohttp~=3.8.5
peewee~=3.16.3